created routes folder and moved server and payments into that folder
