+++ backend/stripe/routes/payment.js
